requirements.txt
requests==2.19.1
beautifulsoup4==4.6.3
schedule==0.5.0
python-dateutil==2.7.5
